+++ SimilarityToolkit.Evaluators/Abstractions/Generic/SimilarityEvaluator.cs
@@ -4,9 +4,9 @@ namespace SimilarityToolkit.Evaluators.Abstractions.Generic
 {
     public abstract class SimilarityEvaluator<T> : SimilarityEvaluator
     {
-        public abstract double EvaluateDistance(T item1, T item2);
+        public abstract decimal EvaluateDistance(T item1, T item2);
 
-        internal override double EvaluateDistance(object item1, object item2)
+        internal override decimal EvaluateDistance(object item1, object item2)
         {
             if (!(item1 is T))
                 throw new InvalidCastException($"{nameof(item1)} must be of type {typeof(T)}.");
