+++ SimilarityToolkit.Evaluators/Abstractions/SimilarityEvaluator.cs
@@ -4,7 +4,7 @@ namespace SimilarityToolkit.Evaluators.Abstractions
 {
     public abstract class SimilarityEvaluator
     {
-        internal abstract double EvaluateDistance(object item1, object item2);
+        internal abstract decimal EvaluateDistance(object item1, object item2);
 
         internal abstract Type EvaluatedType { get; }
     }
