+++ SimilarityToolkit.Evaluators/NullableDoubleSimilarityEvaluator.cs
@@ -5,9 +5,9 @@ namespace SimilarityToolkit.Evaluators
 {
     public class NullableDoubleSimilarityEvaluator : SimilarityEvaluator<double?>
     {
-        public override double EvaluateDistance(double? item1, double? item2)
+        public override decimal EvaluateDistance(double? item1, double? item2)
         {
-            return Math.Abs((item1 ?? 0.0) - (item2 ?? 0.0));
+            return Math.Abs((decimal)(item1 ?? 0.0) - (decimal)(item2 ?? 0.0));
         }
     }
 }
